tjs/hackathon/DocClassificationAI/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
openai>=1.12.0
groq>=0.4.2
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.6.0
aiofiles>=23.2.1
boto3>=1.34.0
botocore>=1.34.0
anthropic>=0.25.0
requests>=2.31.0
numpy>=1.24.0